IOS-3163: Refactor SeedPhraseManager and InputProcessor. Fix review comments
